prompt-to-json-main/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pyjwt
httpx
python-multipart
python-dotenv
passlib[bcrypt]==1.7.4
pydantic-settings
bcrypt==4.0.1
supabase
prometheus-fastapi-instrumentator
sentry-sdk[fastapi]
requests
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1
transformers
trl
accelerate
datasets
peft
gymnasium
stable-baselines3
psycopg2-binary
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
prefect>=2.0.0
PyPDF2>=3.0.0
psycopg2-binary
redis>=4.0.0
psutil>=5.8.0
pdfplumber>=0.7.0
prefect-slack>=0.1.0
prefect-email>=0.1.0
